Split up tests by different files.
